Update README.md
NaijaSDGs-CLICKMEDICAL.
Faced with growing immortality rates in underdeveloped countries due to the lack of rapid support for patients in hospitals in the face of difficulties in purchasing medication for people with motor problems, we decided to provide a solution an online parapharmacy platform.
Indeed, this digital pharmacy will serve as an online drug purchase platform with the possibility of deliveries. It will not only be used to innovate the health system and promote the digitization of professions linked to medicine in Côte d'Ivoire.